+++ app/src/main/AndroidManifest.xml
@@ -9,19 +9,18 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".DineOut"></activity>
-        <activity android:name=".DineIn" />
-        <activity android:name=".MenuView" />
-        <activity android:name=".MainActivity">
-
-        </activity>
-        <activity android:name=".SecondActivity">
+        <activity android:name=".HomeActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".DineOut" />
+        <activity android:name=".DineIn" />
+        <activity android:name=".MenuView" />
+        <activity android:name=".MainActivity"></activity>
+        <activity android:name=".SecondActivity"></activity>
     </application>
 
 </manifest>
