+++ app/src/main/res/layout/activity_second.xml
@@ -91,27 +91,7 @@
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintRight_toRightOf="parent" />
 
-        <Button
-            android:id="@+id/dine_in"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Dine In"
-            android:onClick="onDineInClick"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/recycler_view"
-            app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintLeft_toLeftOf="parent" />
 
-        <Button
-            android:id="@+id/dine_out"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="10dp"
-            android:text="Dine Out"
-            android:onClick="onDineOutClick"
-            app:layout_constraintTop_toBottomOf="@id/dine_in"
-            app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintLeft_toLeftOf="parent" />
 
 
     </androidx.constraintlayout.widget.ConstraintLayout>
